--- a/netsim-async/Cargo.toml
+++ b/netsim-async/Cargo.toml
@@ -15,6 +15,7 @@ tokio = { version = "1.35.1", features = ["sync", "time", "rt", "macros"] }
 
 [dev-dependencies]
 clap = { version = "4.5.1", features = ["derive"] }
+console-subscriber = "0.2.0"
 rand = "0.8.5"
 tokio = { version = "1.35.1", features = [
     "sync",