These examples allow to show how the Mux works under load.

So far, we haven't done any optimisation work and it shows the following:

## Async

The `netsim-async` has an additional latency of 1 to 2 ms on Apple M2. This is the expected value because of how we wait for the next values to be selected. We use [`tokio::time::sleep`] to wait for when the next message is due: https://github.com/input-output-hk/ce-netsim/blob/5049e020a8a0a6e7c6cbd37f07a0e93928485fc1/netsim-async/src/sim_context.rs#L178-L192

However as documented, `tokio::time::sleep` has a granularity of 1 to 2ms depending on the system.

## non async

The current latency will be around 200milliseconds (the hardcoded loop): https://github.com/input-output-hk/ce-netsim/blob/5049e020a8a0a6e7c6cbd37f07a0e93928485fc1/netsim/src/sim_context.rs#L195

However tweaking with this value will allow to change the latency, I managed to locally set it to 200micro-seconds and this dropped the additional latency to ~200microseconds in average.

# outcome

As consequence we will need to rework the `netsim-core` to take the thread wake-up strategy:

1. `netsim`: allow the `sleep` to be configured to a preferred value (with a default to 1ms maybe) #23
2. `netsim-async` and `netsim`: share the same core and remove the dependency on [`tokio::time::sleep`] for the Mux loop. #22 

[`tokio::time::sleep`]: https://docs.rs/tokio/latest/tokio/time/fn.sleep.html